+++ pkg/analyzer/test/src/diagnostics/unused_result_test.dart
@@ -9,12 +9,12 @@ import '../dart/resolution/context_collection_resolution.dart';
 
 main() {
   defineReflectiveSuite(() {
-    defineReflectiveTests(UseResultTest);
+    defineReflectiveTests(UnusedResultTest);
   });
 }
 
 @reflectiveTest
-class UseResultTest extends PubPackageResolutionTest {
+class UnusedResultTest extends PubPackageResolutionTest {
   @override
   void setUp() {
     super.setUp();
@@ -114,7 +114,7 @@ void main() {
   A().foo;
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 100, 7),
+      error(HintCode.UNUSED_RESULT, 104, 3),
     ]);
   }
 
@@ -207,7 +207,7 @@ void main() {
   A().foo;
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 91, 7),
+      error(HintCode.UNUSED_RESULT, 95, 3),
     ]);
   }
 
@@ -224,7 +224,7 @@ void f(bool b) {
   b ? A().foo : 0;
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 98, 7),
+      error(HintCode.UNUSED_RESULT, 102, 3),
     ]);
   }
 
@@ -241,7 +241,7 @@ void f(bool b) {
   b ? (A().foo) : 0;
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 99, 7),
+      error(HintCode.UNUSED_RESULT, 103, 3),
     ]);
   }
 
@@ -262,7 +262,7 @@ void main() {
   });
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 126, 7),
+      error(HintCode.UNUSED_RESULT, 130, 3),
     ]);
   }
 
@@ -346,6 +346,23 @@ void main() {
 ''');
   }
 
+  test_field_static_result_unassigned() async {
+    await assertErrorsInCode(r'''
+import 'package:meta/meta.dart';
+
+class A {
+  @useResult
+  static int foo = 0;
+}
+
+void main() {
+  A.foo;
+}
+''', [
+      error(HintCode.UNUSED_RESULT, 100, 3),
+    ]);
+  }
+
   test_getter_result_passed() async {
     await assertNoErrorsInCode(r'''
 import 'package:meta/meta.dart';
@@ -419,7 +436,9 @@ class A {
 void main() {
   A().foo;
 }
-''', [error(HintCode.UNUSED_RESULT, 96, 7)]);
+''', [
+      error(HintCode.UNUSED_RESULT, 100, 3),
+    ]);
   }
 
   test_method_result_assigned() async {
@@ -472,12 +491,10 @@ int f2() {
   test_method_result_targetedMethod() async {
     await assertNoErrorsInCode(r'''
 import 'package:meta/meta.dart';
-
 class A {
   @useResult
   String foo() => '';
 }
-
 void main() {
   A().foo().toString(); // OK
 }
@@ -512,7 +529,27 @@ void main() {
   A().foo();
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 94, 9),
+      error(HintCode.UNUSED_RESULT, 98, 3),
+    ]);
+  }
+
+  test_method_result_unassigned_cascade() async {
+    await assertErrorsInCode(r'''
+import 'package:meta/meta.dart';
+
+class C {
+  @useResult
+  C m1() => throw '';
+  
+  C m2() => throw '';
+  
+  void m3() {
+    m2()..m1();
+  }
+}
+''', [
+      error(HintCode.UNUSED_RESULT, 131, 2,
+          messageContains: "'m1' should be used."),
     ]);
   }
 
@@ -525,6 +562,20 @@ int foo() => 0;
 
 void main() {
   var x = foo(); // OK
+  print(x);
+}
+''');
+  }
+
+  test_topLevelFunction_result_assigned_cascade() async {
+    await assertNoErrorsInCode(r'''
+import 'package:meta/meta.dart';
+
+@useResult
+int foo() => 0;
+
+void main() {
+  var x = foo()..toString(); // OK
   print(x);
 }
 ''');
@@ -569,7 +620,6 @@ void main() {
 ''');
   }
 
-  // todo(pq):implement visitExpressionStatement?
   test_topLevelFunction_result_unassigned() async {
     await assertErrorsInCode(r'''
 import 'package:meta/meta.dart';
@@ -585,7 +635,22 @@ void main() {
   baz(); // OK
 }
 ''', [
-      error(HintCode.UNUSED_RESULT, 108, 5),
+      error(HintCode.UNUSED_RESULT, 108, 3),
+    ]);
+  }
+
+  test_topLevelFunction_result_unassigned_cascade() async {
+    await assertErrorsInCode(r'''
+import 'package:meta/meta.dart';
+
+@useResult
+int foo() => 0;
+
+void main() {
+  foo()..toString();
+}
+''', [
+      error(HintCode.UNUSED_RESULT, 78, 3),
     ]);
   }