+++ pkg/vm_service/example/vm_service_assert.dart
@@ -159,6 +159,7 @@ String assertEventKind(String obj) {
   if (obj == "ServiceUnregistered") return obj;
   if (obj == "TimelineEvents") return obj;
   if (obj == "TimelineStreamSubscriptionsUpdate") return obj;
+  if (obj == "UserTagChanged") return obj;
   if (obj == "VMFlagUpdate") return obj;
   if (obj == "VMUpdate") return obj;
   if (obj == "WriteEvent") return obj;